#606 Move webpack plugins to wdk
