+++ client/package.json
@@ -25,7 +25,7 @@
     "@dbetka/sass-watcher": "^0.1.5",
     "@dbetka/utils": "^0.1.3",
     "@dbetka/vue-material-icons": "^0.3.2",
-    "@dbetka/wdk": "^0.1.9",
+    "@dbetka/wdk": "^0.2.2",
     "@egjs/flicking-plugins": "^4.2.2",
     "@egjs/vue3-flicking": "^4.4.1",
     "@vuelidate/core": "^2.0.0-alpha.32",
