+++ client/webpack/configs/image-config.js
